requirements.txt
fastapi
uvicorn
pinecone-client
faiss-cpu
openai
langchain
pydantic
python-multipart
aiohttp
pdfplumber
python-docx
pytesseract
Pillow
email6
tqdm
httpx
python-dotenv
